refactor: cleanup saml
