--- cmd/api/src/auth/bhsaml/const.go
@@ -1,30 +1,0 @@
-// Copyright 2023 Specter Ops, Inc.
-//
-// Licensed under the Apache License, Version 2.0
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package bhsaml
-
-const (
-	ObjectIDSurname       = "urn:oid:2.5.4.4"
-	ObjectIDPrincipalName = "urn:oid:1.3.6.1.4.1.5923.1.1.1.6"
-	ObjectIDGivenName     = "urn:oid:2.5.4.42"
-	ObjectIDUserID        = "urn:oid:0.9.2342.19200300.100.1.1"
-	ObjectIDEmail         = "urn:oid:0.9.2342.19200300.100.1.3"
-
-	ObjectIDAttributeNameFormat = "urn:oasis:names:tc:SAML:2.0:attrname-format:uri"
-
-	XMLSOAPClaimsEmailAddress = "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress"
-	XMLTypeString             = "xs:string"
-)
