--- cmd/api/src/auth/bhsaml/tracking.go
@@ -1,34 +1,0 @@
-// Copyright 2023 Specter Ops, Inc.
-//
-// Licensed under the Apache License, Version 2.0
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package bhsaml
-
-import (
-	"net/http"
-
-	"github.com/crewjam/saml"
-	"github.com/crewjam/saml/samlsp"
-)
-
-func NewCookieRequestTracker(serviceProvider ServiceProvider) samlsp.RequestTracker {
-	return samlsp.CookieRequestTracker{
-		ServiceProvider: &serviceProvider.ServiceProvider,
-		NamePrefix:      "sso_",
-		Codec:           samlsp.DefaultTrackedRequestCodec(serviceProvider.Options),
-		MaxAge:          saml.MaxIssueDelay,
-		SameSite:        http.SameSiteDefaultMode,
-	}
-}
