--- cmd/api/src/cmd/testidp/samlidp/util.go
@@ -1,60 +1,0 @@
-package samlidp
-
-import (
-	"bytes"
-	"crypto/rand"
-	"encoding/xml"
-	"errors"
-	"io"
-
-	"github.com/crewjam/saml"
-	xrv "github.com/mattermost/xml-roundtrip-validator"
-)
-
-// RandReader is the io.Reader that produces cryptographically random
-// bytes when they are need by the library. The default value is
-// rand.Reader, but it can be replaced for testing.
-var RandReader = rand.Reader
-
-func randomBytes(n int) []byte {
-	rv := make([]byte, n)
-
-	if _, err := io.ReadFull(RandReader, rv); err != nil {
-		panic(err)
-	}
-	return rv
-}
-
-func getSPMetadata(r io.Reader) (spMetadata *saml.EntityDescriptor, err error) {
-	var data []byte
-	if data, err = io.ReadAll(r); err != nil {
-		return nil, err
-	}
-
-	spMetadata = &saml.EntityDescriptor{}
-	if err := xrv.Validate(bytes.NewBuffer(data)); err != nil {
-		return nil, err
-	}
-
-	if err := xml.Unmarshal(data, &spMetadata); err != nil {
-		if err.Error() == "expected element type <EntityDescriptor> but have <EntitiesDescriptor>" {
-			entities := &saml.EntitiesDescriptor{}
-			if err := xml.Unmarshal(data, &entities); err != nil {
-				return nil, err
-			}
-
-			for _, e := range entities.EntityDescriptors {
-				if len(e.SPSSODescriptors) > 0 {
-					return &e, nil
-				}
-			}
-
-			// there were no SPSSODescriptors in the response
-			return nil, errors.New("metadata contained no service provider metadata")
-		}
-
-		return nil, err
-	}
-
-	return spMetadata, nil
-}
